# pkg/docs/swagger.yaml
definitions:
  ginx.PaginatedResp:
    properties:
      count:
        type: integer
      results: {}
    type: object
  ginx.Response:
    properties:
      data: {}
      message:
        type: string
      requestID:
        type: string
    type: object
...
  serializer.CategoryCreateRequest:
    properties:
      name:
        maxLength: 32
        minLength: 1
        type: string
    required:
    - name
    type: object
  serializer.CategoryListResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      updater:
        type: string
    type: object
  serializer.CategoryRetrieveResponse:
    properties:
      createdAt:
        type: string
      creator:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      updater:
        type: string
    type: object
  serializer.CategoryUpdateRequest:
    properties:
      name:
        maxLength: 32
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  serializer.EntryCreateRequest:
    properties:
      categoryID:
        type: integer
      desc:
        type: string
      name:
        maxLength: 32
        minLength: 1
        type: string
      price:
        type: number
    required:
    - categoryID
    - name
    - price
    type: object
  serializer.EntryListResponse:
    properties:
      categoryID:
        type: integer
      categoryName:
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updatedAt:
        type: string
      updater:
        type: string
    type: object
  serializer.EntryRetrieveResponse:
    properties:
      categoryID:
        type: integer
      categoryName:
        type: string
      createdAt:
        type: string
      creator:
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      updatedAt:
        type: string
      updater:
        type: string
    type: object
  serializer.EntryUpdateRequest:
    properties:
      categoryID:
        type: integer
      desc:
        type: string
      name:
        maxLength: 32
        minLength: 1
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
...
  serializer.PeriodicTaskCreateRequest:
    properties:
      args:
        items: {}
        type: array
      cron:
        type: string
      name:
        type: string
    type: object
  serializer.PeriodicTaskListResponse:
    properties:
      args:
        type: string
      creator:
        type: string
      cron:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: 蓝鲸开发者中心 SaaS 后台 API 文档
  title: BkApp API DOC
  version: 1.0.0
paths:
...
  /api/categories:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serializer.CategoryListResponse'
                  type: array
              type: object
      summary: 获取分类列表
      tags:
      - crud
    post:
      parameters:
      - description: 创建分类请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.CategoryCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: 创建分类
      tags:
      - crud
  /api/categories/{id}:
    delete:
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除分类
      tags:
      - crud
    get:
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  $ref: '#/definitions/serializer.CategoryRetrieveResponse'
              type: object
      summary: 获取单个分类
      tags:
      - crud
    put:
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新分类请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.CategoryUpdateRequest'
      responses:
        "204":
          description: No Content
      summary: 更新分类
      tags:
      - crud
...
  /api/entries:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/ginx.PaginatedResp'
                  - properties:
                      results:
                        items:
                          $ref: '#/definitions/serializer.EntryListResponse'
                        type: array
                    type: object
              type: object
      summary: 获取条目列表
      tags:
      - crud
    post:
      parameters:
      - description: 创建条目请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.EntryCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: 创建条目
      tags:
      - crud
  /api/entries/{id}:
    delete:
      parameters:
      - description: 条目 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除条目
      tags:
      - crud
    get:
      parameters:
      - description: 条目 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  $ref: '#/definitions/serializer.EntryRetrieveResponse'
              type: object
      summary: 获取单个条目
      tags:
      - crud
    put:
      parameters:
      - description: 条目 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新条目请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.EntryUpdateRequest'
      responses:
        "204":
          description: No Content
      summary: 更新条目
      tags:
      - crud
...
  /api/periodic-tasks:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/serializer.PeriodicTaskListResponse'
                  type: array
              type: object
      summary: 获取定时任务列表
      tags:
      - async-task
    post:
      parameters:
      - description: 定时任务配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.PeriodicTaskCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ginx.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: 创建定时任务
      tags:
      - async-task
  /api/periodic-tasks/{id}:
    delete:
      parameters:
      - description: 定时任务 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除定时任务
      tags:
      - async-task
  /api/periodic-tasks/{id}/enabled:
    put:
      parameters:
      - description: 定时任务 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 切换定时任务启用状态
      tags:
      - async-task
...
swagger: "2.0"